New app Model, Func, Dto model

+++ internal/models/Group_model.go
@@ -2,7 +2,7 @@ package models
 
 type EinfGroup struct {
 	Id         int
-	Course     int
+	Course     byte
 	Speciality string
 	Groudates  byte
 	Semester   byte
